Moved EmpytResultSet to django.core.exceptions.
This removes the need for some inner imports.
