+++ django/db/models/sql/__init__.py
@@ -1,4 +1,4 @@
-from django.db.models.sql.datastructures import EmptyResultSet
+from django.core.exceptions import EmptyResultSet
 from django.db.models.sql.query import *  # NOQA
 from django.db.models.sql.subqueries import *  # NOQA
 from django.db.models.sql.where import AND, OR
